refactor: update domain model imports and clean up TaskEditorBottomSheetContent layout
